refactor(#3103): extract legacy-regexp cluster from runtime.ts
Slice 2 of the src/runtime.ts (#3103) god-file decomposition. Lift the
Annex B legacy RegExp static-state machinery (RegExp.$1..$9 / $_ / $& /
lastMatch etc.) plus the String.prototype symbol-method reroute for
primitive search values (#3095) verbatim into src/runtime/legacy-regexp.ts.

runtime.ts shrinks 15,025 -> 14,834 LOC (-191). Bodies moved verbatim;
only added lines are the license header, `export` keywords, and one
import block back into runtime.ts. One-directional import (no cycle).
Privates fully contained (0 stray refs remain); five symbols imported back.

runtime.ts is host-side JS, not in the Wasm emit path, so the split cannot
change an emitted byte by construction. Isolated tsc clean; check:loc-budget
green; targeted vitest (matchall, standalone-regex, issue-3014) green.

✓
